Partially enable final static field folding
- Add support for final field folding in
`J9::TransformUtil::transformDirectLoad`.
This method dereferences a static address on the client,
so replace dereferencing with calling a new method of
ResolvedJ9Method, which will make an appropripate
remote call on the server.
Returned dereferenced values are cached to minimize the number of remote
calls.

- Enable folding of initializeStatus field in J9Class. To determine
whether
we can fold this field just need to check if a class is initialized on
the client. We already have a remote method for that, with caching.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>