Partially enable final static field folding
Add support for final field folding in
`J9::TransformUtil::transformDirectLoad`.
This method dereferences a static address on the client,
so replace dereferencing with calling a new method of
ResolvedJ9Method, which will make an appropripate
remote call on the server.
Returned dereferenced values are cached to minimize the number of remote
calls.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>