Handle disabled Known Object Table in invokeFilterArgumentsHandle
Previously, if a filter arguments method handler was compiled,
known object table had to be enabled, otherwise the compiler would
crash.
However, Known Object Table is only used to improve an existing symbol
reference,
it's not required for compilation.
This change enables to run with disabled known object table, by
using a list of filters, instead of a list of filter indices, to check
if a null filter is used.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>