Reduce number of VM_jitFields(Statics)AreSame messages
Currently, `VM_jitFieldsAreSame` is the most frequent message
in AcmeAir.
Looking at the implementation of `jitFieldsAreIdentical`, it
considers fields to be identical when they belong to the same
Java class and contain the same value.
We can check if they belong to the same class on the server,
and if cp indices are the same, then fields have to be identical.
This improvement should significantly reduce the number of remote
messages.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>