Support new invokedynamic and invokehandle bytecode handlers for JITServer
This enables JITServer support for #10607.
Parts of the new Method Handle code require holding
VM access in the JIT, so to make it work on JITServer,
all such code has been wrapped in front-end calls.
When executed on the server, these front-end queries will
send a message to the client, which will do the required
work while holding VM access and send the result to the server.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>